add relevant components and routes for Calendar Page
