+++ src/components/ChoosedDay.jsx
@@ -1,0 +1,5 @@
+const ChoosedDay = () => {
+  return <div>ChoosedDay</div>;
+};
+
+export default ChoosedDay;
