+++ src/components/MonthCalendarHead/MonthCalendarHead.jsx
@@ -1,7 +1,7 @@
 import moment from 'moment';
-import { WeekDaysWrapper, WeekDay } from './CalendarWeekDays.styled';
+import { WeekDaysWrapper, WeekDay } from './MonthCalendarHead.styled';
 
-const CalendarWeekDays = ({ startDay }) => {
+const MonthCalendarHead = ({ startDay }) => {
   const day = startDay.clone().subtract(1, 'day');
   const daysOfWeek = [...Array(7)].map(() => day.add(1, 'day').clone());
 
@@ -9,7 +9,7 @@ const CalendarWeekDays = ({ startDay }) => {
     <WeekDaysWrapper>
       {daysOfWeek.map((dayItem, idx) => (
         <li key={dayItem.unix()}>
-          <WeekDay isWeekendDay={dayItem.day() === 6 || dayItem.day() === 0}>
+          <WeekDay $isWeekendDay={dayItem.day() === 6 || dayItem.day() === 0}>
             {moment()
               .day(idx + 1)
               .format('ddd')}
@@ -20,4 +20,4 @@ const CalendarWeekDays = ({ startDay }) => {
   );
 };
 
-export default CalendarWeekDays;
+export default MonthCalendarHead;
