+++ src/components/PeriodPaginator/PeriodPaginator.jsx
@@ -1,3 +1,22 @@
-export default function PeriodPaginator() {
-  return <p>Period Paginator</p>;
+import {
+  BtnNext,
+  BtnPrev,
+  PaginatorWrapper,
+  Period,
+} from './PeriodPaginator.styled';
+
+export default function PeriodPaginator({
+  currentMonth,
+  prevMonthHandler,
+  nextMonthHandler,
+}) {
+  return (
+    <PaginatorWrapper>
+      <Period>{currentMonth.format('MMMM YYYY')}</Period>
+      <div>
+        <BtnPrev onClick={prevMonthHandler}>&lt;</BtnPrev>
+        <BtnNext onClick={nextMonthHandler}>&gt;</BtnNext>
+      </div>
+    </PaginatorWrapper>
+  );
 }
