+++ src/components/PeriodTypeSelect/PeriodTypeSelect.jsx
@@ -1,0 +1,12 @@
+import { BtnDay, BtnMonth } from './PeriodTypeSelect.styled';
+
+const PeriodTypeSelect = () => {
+  return (
+    <div>
+      <BtnMonth>Month</BtnMonth>
+      <BtnDay>Day</BtnDay>
+    </div>
+  );
+};
+
+export default PeriodTypeSelect;
